Hardcode like svp
To not be dependent of leng, need to search on there thing than string. On Firefox can search the definition of the svg but Chrome it's hidden. So hardcoded for the moment was teh best solution
The like button in comment have the svg code too, but the user have to scroll to comment to start load of this section so we could not get the pattern before the user scroll : not relevant