Fix yt playlist
Youtube changed the events triggered when browsing on the website. Now
a check of the url need to be done on each update. Else addon is
triggered on home page etc. and wait until end of time. When going on
a new video, the 'home' script is still waiting and throw an error
to killing the right script (block multiple run).